requirements.txt
django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
dj-rest-auth==2.2.5
django-allauth==0.51.0